refactor(ui): reduce ProcessTree.tsx cognitive complexity via pure-helper extraction [no-behavior-change]
consolidation-pass paydown on the repo's highest-complexity file (cog ~253). Move the pure module-level functions + the two shared types to a new pure sibling ProcessTree.helpers.ts (no React, no d3), lift the three useMemo bodies into pure build* helpers (buildPreservedIds / buildQueryFilterIds / buildVisibleRoots), and de-nest renderTree's inline branch callbacks into small pure helpers (node dot/label/chevron class+fill+text, computeLayoutBounds). renderTree itself (the d3 DOM mutation) stays in place.

Behavior-preserving: verbatim moves / closure-to-parameter only. No hook order changed, no useMemo/useEffect dependency array changed, no D3 enter/update/exit/join logic changed, no component state shape changed. Verified: eslint + tsc clean, ProcessTree.test.tsx (14) green, full vitest suite (597) green.
